Update Player::Loader with update_all_modes
- remove loader & license field from site
- remove cdn_up_to_date field from site

+++ app/models/player/loader.rb
@@ -1,23 +1,26 @@
 # TODO
-# - per site
 # - include component relation with good version depending on the site.player_mode (stabe, beta, alpha) & app component
 
 require 'tempfile'
+require 'digest/md5'
 require_dependency 'cdn'
 
 class Player::Loader < Struct.new(:site, :mode, :file)
+  MODES = %w[stable beta alpha]
   delegate :token, :player_mode, to: :site
 
-  def self.update!(site_id, mode)
+  def self.update_all_modes!(site_id)
     site = Site.find(site_id)
-    loader = new(site, mode)
-    loader.upload!
-  end
-
-  def self.delete!(site_id)
-    site = Site.find(site_id)
-    loader = new(site, mode)
-    loader.delete!
+    modes_needed = site_loader_modes(site)
+    changed = []
+    MODES.each do |mode|
+      if modes_needed.include?(mode)
+        changed << new(site, mode).upload!
+      else
+        new(site, mode).delete!
+      end
+    end
+    site.touch(:loaders_updated_at) if changed.any?
   end
 
   def initialize(*args)
@@ -26,34 +29,67 @@ class Player::Loader < Struct.new(:site, :mode, :file)
   end
 
   def upload!
-    File.open(file) do |f|
-      data = f.read
-      mode_config[:destinations].each do |destination|
-        s3_put_object(
-          destination[:bucket],
-          destination[:path],
-          data
-        )
+    if changed?
+      File.open(file) do |f|
+        data = f.read
+        mode_config[:destinations].each do |destination|
+          s3_put_object(
+            destination[:bucket],
+            destination[:path],
+            data
+          )
+        end
       end
+      purge_cdn
+      true
     end
-    purge_cdn
   end
 
   def delete!
-    mode_config[:destinations].each do |destination|
-      S3.fog_connection.delete_object(
-        destination[:bucket],
-        destination[:path]
-      )
+    if present?
+      mode_config[:destinations].each do |destination|
+        S3.fog_connection.delete_object(
+          destination[:bucket],
+          destination[:path]
+        )
+      end
+      purge_cdn
+      true
     end
-    purge_cdn
   end
 
   def components_path
     [] # TODO Thibaud
   end
 
+  # loader already present on S3?
+  def present?
+    mode_config[:destinations].all? do |destination|
+      s3_headers(
+        destination[:bucket],
+        destination[:path]
+      ).present?
+    end
+  end
+
+  # loader different that the one already present on S3?
+  def changed?
+    md5 != uploaded_md5
+  end
+
 private
+
+  def self.site_loader_modes(site)
+    if site.state == 'active'
+      case site.player_mode
+      when 'stable'; %w[stable]
+      when 'beta'; %w[stable beta]
+      when 'alpha'; %w[stable beta alpha]
+      end
+    else
+      []
+    end
+  end
 
   def generate_file
     template_path = Rails.root.join("app/templates/player/#{mode_config[:template]}")
@@ -61,6 +97,18 @@ private
     self.file = Tempfile.new("l-#{site.token}.js", "#{Rails.root}/tmp")
     self.file.print template.result(binding)
     self.file.flush
+  end
+
+  def md5
+    File.open(file) { |f| Digest::MD5.base64digest(f.read) }
+  end
+
+  def uploaded_md5
+    destination = mode_config[:destinations].first
+    s3_headers(
+      destination[:bucket],
+      destination[:path]
+    )['Content-MD5']
   end
 
   def s3_put_object(bucket, path, data)
@@ -71,9 +119,19 @@ private
       {
         'Cache-Control' => 'max-age=120, public', # 2 minutes
         'Content-Type'  => 'text/javascript',
+        'Content-MD5'   => md5,
         'x-amz-acl'     => 'public-read'
       }
     )
+  end
+
+  def s3_headers(bucket, path)
+    S3.fog_connection.head_object(
+      bucket,
+      path
+    ).headers
+  rescue Excon::Errors::NotFound
+    {}
   end
 
   def purge_cdn
